Only delete category product associtions during product import if at least one category is assigned in 'replace' mode

+++ src/app/code/community/AvS/FastSimpleImport/Model/Import/Entity/Product.php
@@ -1326,13 +1326,16 @@ class AvS_FastSimpleImport_Model_Import_Entity_Product extends AvS_FastSimpleImp
 
             foreach ($categoriesData as $delSku => $categories) {
                 $productId      = $this->_newSku[$delSku]['entity_id'];
-                $delProductId[] = $productId;
 
                 foreach ($categories as $categoryId => $position) {
+                    $delProductId[$productId] = $productId;
                     $categoriesIn[] = array('product_id' => $productId, 'category_id' => $categoryId, 'position' => (int) $position);
                 }
             }
-            if (Mage_ImportExport_Model_Import::BEHAVIOR_APPEND != $this->getBehavior()) {
+            if (Mage_ImportExport_Model_Import::BEHAVIOR_APPEND != $this->getBehavior() && sizeof($delProductId)) {
+                print_r($categoriesData);
+                print_r($delProductId);
+                die();
                 $this->_connection->delete(
                     $tableName,
                     $this->_connection->quoteInto('product_id IN (?)', $delProductId)
